Repairing hakyll.hs (it has never been committed)
I hope, it is my biggest fail ever. The Hakyll source for a webpage
has never been committed and I have no access to the machine where
I've wrote that originally.

Recreating it from scratch now. A good opportunity to redesign a page :)

News and Downloads are broken and TBD. The front page is under construction
too (removed that sliderman, we do not need that).